Allow turning folder-scoped provider on/off for a folder
