+++ src/main/java/io/jenkins/plugins/credentials/secretsmanager/folders/supplier/FolderCredentialsSupplier.java
@@ -24,9 +24,11 @@ public class FolderCredentialsSupplier {
 
             // the Jenkins user may not have configured this plugin on the folder
             if (folderPluginConfiguration != null) {
-                final var pluginConfiguration = folderPluginConfiguration.build();
+                final var ourPluginConfiguration = folderPluginConfiguration.getPluginConfiguration();
 
-                return credentialsSupplier.get(pluginConfiguration);
+                if (ourPluginConfiguration != null) {
+                    return credentialsSupplier.get(ourPluginConfiguration.build());
+                }
             }
         }
