+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/folders/config/FolderPluginConfiguration/config.jelly
@@ -1,20 +1,6 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
     <f:section title="${%awsSecretsManagerCredentialsProvider}">
-        <f:entry title="${%cache}" field="cache">
-            <f:checkbox default="true" />
-        </f:entry>
-        <f:block>
-            <div class="section-header" style="border-bottom: none">${%client}</div>
-        </f:block>
-        <f:property field="client" />
-        <f:block>
-            <div class="section-header" style="border-bottom: none">${%listSecrets}</div>
-        </f:block>
-        <f:property field="listSecrets" />
-        <f:block>
-            <div class="section-header" style="border-bottom: none">${%transformations}</div>
-        </f:block>
-        <f:property field="transformations" />
+        <f:optionalProperty field="pluginConfiguration"  title="${%enabled}"/>
     </f:section>
 </j:jelly>
