+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/folders/config/FolderPluginConfiguration/config_en.properties
@@ -1,5 +1,2 @@
 awsSecretsManagerCredentialsProvider = AWS Secrets Manager Credentials Provider (folder-scoped)
-cache=Cache
-client = Client
-listSecrets = ListSecrets
-transformations = Transformations
+enabled = Enabled
