+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/folders/config/OurPluginConfiguration/config_en.properties
@@ -1,0 +1,4 @@
+cache = Cache
+client = Client
+listSecrets = ListSecrets
+transformations = Transformations
